fix(secrets): keep node:sqlite out of the secrets command path
`monitors --json` asserts a clean stderr, and it started failing on this branch
while passing on main. The cause was mine and worth stating plainly:
list-filter.ts imported `parseTimeFilter` from `session/discover.js`, and
discover.ts imports `../sqlite.js`. Reaching through it for one duration parser
pulled node:sqlite into the `agents secrets` module graph, so Node printed

  (node:730848) ExperimentalWarning: SQLite is an experimental feature ...

on stderr for every CLI invocation that registers the secrets commands — i.e.
all of them. Invisible in a command that already touches the session DB, a
broken contract in one that does not.

Fixed at the source rather than worked around: `parseTimeFilter` moves to
session/relative-time.ts, which has no imports at all, and discover.ts
re-exports it so `sessions`, `cost`, `teams`, and `output` are untouched.

Verified: a `secrets list` run now emits zero lines on stderr, and
monitors.test.ts passes again.
