+++ apps/cli/src/lib/secrets/list-filter.ts
@@ -19,7 +19,10 @@
  */
 
 import { parseCommaSeparatedList } from '../../commands/utils.js';
-import { parseTimeFilter } from '../session/discover.js';
+// From the leaf module, NOT discover.js — the latter imports `../sqlite.js`, so
+// reaching through it for one parser would pull node:sqlite into every `agents
+// secrets` invocation and print Node's SQLite ExperimentalWarning on stderr.
+import { parseTimeFilter } from '../session/relative-time.js';
 import {
   describeBundle,
   bundlePolicy,
